Moved file handing into module
